# acteos_lifeos_romania_v2_1/contracts/openapi.yaml
openapi: 3.1.1
info:
  title: ActeOS API
  version: 1.1.0
  description: API contract pentru ActeOS / LifeOS România v2.1. Discovery este intent-first; regulile administrative critice
    sunt determinate de rulesets aprobate, nu de LLM runtime.
servers:
- url: https://api.example.invalid
  description: placeholder; configure per environment
...
paths:
...
  /v1/intents:
    get:
      summary: Caută sau listează intenturi canonice publicate
      operationId: listIntents
      tags:
      - Intents
      parameters:
      - name: query
        in: query
        schema:
          type:
          - string
          - 'null'
          maxLength: 200
      - name: category_id
        in: query
        schema:
          type:
          - string
          - 'null'
      - name: jurisdiction_id
        in: query
        schema:
          type:
          - string
          - 'null'
      - name: cursor
        in: query
        schema:
          type:
          - string
          - 'null'
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntentListResponse'
        '400': *id001
        '401': *id002
        '403': *id003
        '404': *id004
        '409': *id005
        '422': *id006
        '429': *id007
        '500': *id008
      security: []
...
  /v1/cases:
    post:
      summary: Creează un caz/eveniment al utilizatorului
      operationId: createCase
      tags:
      - Cases
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400': *id009
        '401': *id010
        '403': *id011
        '404': *id012
        '409': *id013
        '422': *id014
        '429': *id015
        '500': *id016
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      security:
      - anonymousToken: []
      - bearerAuth: []
  /v1/cases/{case_id}:
    get:
      summary: Returnează cazul
      operationId: getCase
      tags:
      - Cases
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400': *id009
        '401': *id010
        '403': *id011
        '404': *id012
        '409': *id013
        '422': *id014
        '429': *id015
        '500': *id016
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      security:
      - anonymousToken: []
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    anonymousToken:
      type: apiKey
      in: header
      name: X-Anonymous-Token
  schemas:
...
    PageInfo:
      type: object
      required:
      - has_more
      properties:
        next_cursor:
          type:
          - string
          - 'null'
        has_more:
          type: boolean
...
    CreateCaseRequest:
      type: object
      required:
      - intent_type_id
      - reference_date
      - timezone
      - jurisdiction_path
      properties:
        intent_type_id:
          type: string
          pattern: ^ro\.intent\.[a-z0-9_.-]+$
        reference_date:
          type: string
          format: date
        timezone:
          type: string
          const: Europe/Bucharest
        jurisdiction_path:
          type: array
          minItems: 2
          items:
            type: string
        event_type_id:
          type:
          - string
          - 'null'
          pattern: ^ro\.life\.[a-z0-9_.-]+$
        user_id:
          type:
          - string
          - 'null'
        installation_id:
          type:
          - string
          - 'null'
        discovery_source:
          type:
          - string
          - 'null'
          enum:
          - search
          - category
          - quick_action
          - related_intent
          - deep_link
          - legacy_event
          - null
        facts:
          type: object
          additionalProperties: true
    CaseResponse:
      type: object
      required:
      - id
      - intent_type_id
      - event_type_id
      - reference_date
      - timezone
      - jurisdiction_path
      - status
      - version
      - facts_hash
      - engine_version
      - ruleset_version
      - events
      - resolution_trace
      - created_at
      properties:
        id:
          type: string
          format: uuid
        intent_type_id:
          type: string
        event_type_id:
          type: string
        reference_date:
          type: string
          format: date
        timezone:
          type: string
        jurisdiction_path:
          type: array
          items:
            type: string
        status:
          type: string
        version:
          type: integer
        facts_hash:
          type: string
        engine_version:
          type: string
        ruleset_version:
          type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventResolution'
        resolution_trace:
          $ref: '#/components/schemas/ResolutionTrace'
        created_at:
          type: string
          format: date-time
    EventResolution:
      type: object
      required:
      - event_type_id
      - status
      - included_steps
      - requirements
      - obligations
      - channels
      - advice
      - warnings
      - confirmations
      - conflicts
      - blocks
      - deadlines
      - overdue_steps
      - child_events
      - missing_facts
      - overridden_rules
      - fired_rule_ids
      - excluded_rule_ids
      properties:
        event_type_id:
          type: string
        status:
          type: string
        included_steps:
          type: array
          items:
            type: string
        requirements:
          type: array
          items:
            type: string
        obligations:
          type: object
          additionalProperties: true
        channels:
          type: array
          items:
            type: string
        deadlines:
          type: object
          additionalProperties: true
        advice:
          type: array
          items:
            type: object
            additionalProperties: true
        warnings:
          type: array
          items:
            type: object
            additionalProperties: true
        confirmations:
          type: array
          items:
            type: string
        conflicts:
          type: array
          items:
            type: string
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
        overdue_steps:
          type: array
          items:
            type: string
        child_events:
          type: array
          items:
            type: string
        missing_facts:
          type: array
          items:
            type: string
        overridden_rules:
          type: array
          items:
            type: string
        fired_rule_ids:
          type: array
          items:
            type: string
        excluded_rule_ids:
          type: array
          items:
            type: string
    ResolutionTrace:
      type: object
      required:
      - facts_hash
      - engine_version
      - included_rule_ids
      - excluded_rule_ids
      - ruleset_version
      - root_event_type_id
      - event_type_ids
      properties:
        facts_hash:
          type: string
        engine_version:
          type: string
        included_rule_ids:
          type: array
          items:
            type: string
        excluded_rule_ids:
          type: array
          items:
            type: string
        ruleset_version:
          type: string
        reference_date:
          type:
          - string
          - 'null'
        root_event_type_id:
          type: string
        event_type_ids:
          type: array
          items:
            type: string
        has_unknown_events:
          type: boolean
          default: false
        intent_type_id:
          type:
          - string
          - 'null'
...
    IntentSummary:
      type: object
      required:
      - id
      - category_id
      - kind
      - title_ro
      - outcome_ro
      - production_status
      - available
      properties:
        id:
          type: string
        category_id:
          type: string
        kind:
          type: string
          enum:
          - direct_goal
          - bundle_goal
        title_ro:
          type: string
        outcome_ro:
          type: string
        production_status:
          type: string
        available:
          type: boolean
        availability_message_ro:
          type:
          - string
          - 'null'
        release_wave:
          type: string
    IntentListResponse:
      type: object
      required:
      - items
      - catalog_version
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IntentSummary'
        page:
          $ref: '#/components/schemas/PageInfo'
        catalog_version:
          type: string
...